Extract requireFirst to helpers module
