+++ bin/pa11y.js
@@ -8,6 +8,7 @@ const path = require('path');
 const pkg = require('../package.json');
 const {program} = require('commander');
 const pa11y = require('../lib/pa11y');
+const {requireFirst} = require('../lib/helpers');
 const semver = require('semver');
 
 const programOptions = program.opts();
@@ -250,26 +251,6 @@ function checkReporterCompatibility(reporterName, reporterSupportString, pa11yVe
 		console.error(`Reporter Support: ${reporterSupportString}`);
 		console.error(`Pa11y Version:    ${pa11yVersion}`);
 		process.exit(1);
-	}
-}
-
-/**
- * Traverses a number of directories trying to load a config file from them
- * @param {String[]} stack - A list of directories
- * @param {Object} defaultReturn - The object to return if no config is found
- * @returns {Object} A config object
- */
-function requireFirst(stack, defaultReturn) {
-	if (!stack.length) {
-		return defaultReturn;
-	}
-	try {
-		return require(stack.shift());
-	} catch (error) {
-		if (error.code === 'MODULE_NOT_FOUND') {
-			return requireFirst(stack, defaultReturn);
-		}
-		throw error;
 	}
 }
